Fix add_str_to_exc in case of UnicodeError
Adding tests showed there was an error as the `repr()`
representation did not contain the added string.
Fix is to add to the `reason` member if it exists.
This implies a bit of refactoring to remove duplicated
code.
On the long run, it could be a good idea to run the test
on all subclasses of Exception.