Refactor tests related to `add_string_to_exc`
As I was getting a bit annoyed of discovering
un-handled edge cases for `add_string_to_exc`
every now and then, I wanted to have a way to
test the function for every single Exception
type once and for all.
The corresponding tests are not commited
because the way it worked was ugly, it didn't
show any new issues while it artificially
boosted the number of tests runs.

However, while doing so, I took this chance to
refactor things and if I ever want to, I can
easily have the tests runs for all exceptions.