Handle cases where traceback is None
Traceback may be None in some cases and this
needs to be handled properly.
When this happens, we may not be able to
suggest as much as when we do have the traceback.

Taking this chance to add a bit of documentation.
(Even though I'd like to rewrite this soon)