Check that only one regexp matches
Not effective yet but it could be a good idea
to match that a single regexp matches the
different messages tested in unit tests.
This is deactivated by default because 2 regexps
are ambiguous but this should be "easy" to fix
(a bit painful because of the various Python
versions/interpreters used).

Also, take this change to add the missing ^/$ in
recent regexp.