Make regexp unambiguous
It is better to ensure in unit tests that
a single test message cannot be matched by
multiple regular expressions.
Also, this leadis to splitting some test
cases based on the Python version/interpreter
and in particular mess with Python 2.7.
This also leads to having the same regexp
used for both AttributeError and TypeError
but fortunately, the register decorator
deals with this nicely.

Corresponds to the squash/merge of the
various commits on dev branch called
"unambiguous_regexp".