Perform proper cleanup in PostMortemTest #42
PostMortemTest leaves a few attributes set that
should not be. This causes side-effects in other
test suites (didyoumean_sugg_tests).

In order to do so, cleanup is performed and
assertions are added to check that everything
is clean at the beginning and at the end.

It is interesting to see that in:
 commit b2bb75297450d64d66e4ecb65f8374f3f98aacd5
 Date:   Tue Jun 30 19:17:59 2015 +0100
I probably had a cleanup issue but solved it
the wrong way by doing the cleanup at the
beginning rather than at the end...