Disable pep8 checks disabled by default
From the doc:
""""
In the default configuration, the checks E121, E123, E126, E133, E226, E241, E242, E704, W503, W504 and W505 are ignored because they are not rules unanimously accepted, and PEP 8 does not enforce them. Please note that if the option –ignore=errors is used, the default configuration will be overridden and ignore only the check(s) you skip.
"""".

Thus trying to disable warnings in previous
commit actually also enables a few warnings.
This should fix the issue.