Handle 'buffer' suggested as attrs from non-local objects
Following the implementation of non-local objects,
a new problem was revealed by Jenkins: for the NameError
related to buffer, we may end up with various suggs
such as 'obj.buffer' where obj is an object in the
enclosing scope highly dependent on the context
(testing framework, Python version, etc).

These can be hard to test so a dirty way to handle
these have been added.

Note: the suggestion variable is not called "sugg_buffer"
or anything similar so that it does not appear in the
list of suggestions.