Handle non-interesting context-related suggestions for NameError
Follow-up of the two previous commits.
Following the implementation of non-local objects,
a new problem was revealed by Jenkins: a few suggestions
for NameError are highly sensitive the objects in frame
and in particular, in the enclosing scope which highly
dependent on the context (testing framework, Python version, etc).

This makes the definition of the expected suggestions in
unit-tests a bit tricky to handle. This was somehow handled
for buffer but a few other places are still failing:
 - file (Python 3.2)
 - loader (Python 2.6)